feat: handle update jobs and delete jobs

+++ src/views/JobView.vue
@@ -29,6 +29,8 @@
 
     const deleteJob = async () => {
         try {
+            const confirm = window.confirm('Are you sure you want to delete this job?');
+            if (!confirm) return;
             await axios.delete(`/api/jobs/${jobId}`);
             toast.success('Job deleted successfully');
             router.push('/jobs');
@@ -104,7 +106,7 @@
             <div class="bg-white p-6 rounded-lg shadow-md mt-6">
               <h3 class="text-xl font-bold mb-6">Manage Job</h3>
               <RouterLink
-                to="/jobs/edit"
+                :to="'/jobs/edit/' + jobId"
                 class="bg-green-500 hover:bg-green-600 text-white text-center font-bold py-2 px-4 rounded-full w-full focus:outline-none focus:shadow-outline mt-4 block"
                 >Edit Job
                 </RouterLink>
